+++ lib/appwrite/enums/project_o_auth_provider_id.rb
@@ -44,8 +44,6 @@ module Appwrite
             YANDEX = 'yandex'
             ZOHO = 'zoho'
             ZOOM = 'zoom'
-            GITHUBIMAGINE = 'githubImagine'
-            GOOGLEIMAGINE = 'googleImagine'
         end
     end
 end