+++ lib/appwrite/enums/region.rb
@@ -1,0 +1,12 @@
+module Appwrite
+    module Enums
+        module Region
+            FRA = 'fra'
+            NYC = 'nyc'
+            SYD = 'syd'
+            SFO = 'sfo'
+            SGP = 'sgp'
+            TOR = 'tor'
+        end
+    end
+end