+++ lib/appwrite/models/policy_deny_free_email.rb
@@ -1,0 +1,32 @@
+#frozen_string_literal: true
+
+module Appwrite
+    module Models
+        class PolicyDenyFreeEmail
+            attr_reader :id
+            attr_reader :enabled
+
+            def initialize(
+                id:,
+                enabled:
+            )
+                @id = id
+                @enabled = enabled
+            end
+
+            def self.from(map:)
+                PolicyDenyFreeEmail.new(
+                    id: map["$id"],
+                    enabled: map["enabled"]
+                )
+            end
+
+            def to_map
+                {
+                    "$id": @id,
+                    "enabled": @enabled
+                }
+            end
+        end
+    end
+end