+++ lib/appwrite/services/functions.rb
@@ -59,12 +59,14 @@ module Appwrite
         # @param [String] provider_branch Production branch for the repo linked to the function.
         # @param [] provider_silent_mode Is the VCS (Version Control System) connection in silent mode for the repo linked to the function? In silent mode, comments will not be made on commits and pull requests.
         # @param [String] provider_root_directory Path to function code in the linked repo.
+        # @param [Array] provider_branches List of branch name patterns to trigger automatic deployments. Supports wildcards. Leave empty to deploy on all branches.
+        # @param [Array] provider_paths List of file path patterns to trigger automatic deployments. Supports wildcards. Leave empty to deploy on all file changes.
         # @param [String] build_specification Build specification for the function deployments.
         # @param [String] runtime_specification Runtime specification for the function executions.
         # @param [Integer] deployment_retention Days to keep non-active deployments before deletion. Value 0 means all deployments will be kept.
         #
         # @return [Function]
-        def create(function_id:, name:, runtime:, execute: nil, events: nil, schedule: nil, timeout: nil, enabled: nil, logging: nil, entrypoint: nil, commands: nil, scopes: nil, installation_id: nil, provider_repository_id: nil, provider_branch: nil, provider_silent_mode: nil, provider_root_directory: nil, build_specification: nil, runtime_specification: nil, deployment_retention: nil)
+        def create(function_id:, name:, runtime:, execute: nil, events: nil, schedule: nil, timeout: nil, enabled: nil, logging: nil, entrypoint: nil, commands: nil, scopes: nil, installation_id: nil, provider_repository_id: nil, provider_branch: nil, provider_silent_mode: nil, provider_root_directory: nil, provider_branches: nil, provider_paths: nil, build_specification: nil, runtime_specification: nil, deployment_retention: nil)
             api_path = '/functions'
 
             if function_id.nil?
@@ -97,6 +99,8 @@ module Appwrite
                 providerBranch: provider_branch,
                 providerSilentMode: provider_silent_mode,
                 providerRootDirectory: provider_root_directory,
+                providerBranches: provider_branches,
+                providerPaths: provider_paths,
                 buildSpecification: build_specification,
                 runtimeSpecification: runtime_specification,
                 deploymentRetention: deployment_retention,
@@ -210,12 +214,14 @@ module Appwrite
         # @param [String] provider_branch Production branch for the repo linked to the function
         # @param [] provider_silent_mode Is the VCS (Version Control System) connection in silent mode for the repo linked to the function? In silent mode, comments will not be made on commits and pull requests.
         # @param [String] provider_root_directory Path to function code in the linked repo.
+        # @param [Array] provider_branches List of branch name patterns to trigger automatic deployments. Supports wildcards. Leave empty to deploy on all branches.
+        # @param [Array] provider_paths List of file path patterns to trigger automatic deployments. Supports wildcards. Leave empty to deploy on all file changes.
         # @param [String] build_specification Build specification for the function deployments.
         # @param [String] runtime_specification Runtime specification for the function executions.
         # @param [Integer] deployment_retention Days to keep non-active deployments before deletion. Value 0 means all deployments will be kept.
         #
         # @return [Function]
-        def update(function_id:, name:, runtime: nil, execute: nil, events: nil, schedule: nil, timeout: nil, enabled: nil, logging: nil, entrypoint: nil, commands: nil, scopes: nil, installation_id: nil, provider_repository_id: nil, provider_branch: nil, provider_silent_mode: nil, provider_root_directory: nil, build_specification: nil, runtime_specification: nil, deployment_retention: nil)
+        def update(function_id:, name:, runtime: nil, execute: nil, events: nil, schedule: nil, timeout: nil, enabled: nil, logging: nil, entrypoint: nil, commands: nil, scopes: nil, installation_id: nil, provider_repository_id: nil, provider_branch: nil, provider_silent_mode: nil, provider_root_directory: nil, provider_branches: nil, provider_paths: nil, build_specification: nil, runtime_specification: nil, deployment_retention: nil)
             api_path = '/functions/{functionId}'
                 .gsub('{functionId}', function_id)
 
@@ -244,6 +250,8 @@ module Appwrite
                 providerBranch: provider_branch,
                 providerSilentMode: provider_silent_mode,
                 providerRootDirectory: provider_root_directory,
+                providerBranches: provider_branches,
+                providerPaths: provider_paths,
                 buildSpecification: build_specification,
                 runtimeSpecification: runtime_specification,
                 deploymentRetention: deployment_retention,